requirements.txt
charset-normalizer==2.0.3
Django==3.2.5
flake8==3.9.2
pip-check-reqs==2.2.2
requests==2.26.0
urllib3==1.26.6
psycopg2-binary==2.9.1
djangorestframework==3.12.4
gunicorn==20.1.0
requests-mock==1.9.3
